tty fixed, refactoring

+++ src/klib/vga.h
@@ -31,8 +31,8 @@ enum vga_color
     COLOR_WHITE = 15,
 };
 
-static const size_t VGA_WIDTH = 80;
-static const size_t VGA_HEIGHT = 25;
+static const uint16_t VGA_WIDTH = 80;
+static const uint16_t VGA_HEIGHT = 25;
 
 static uint16_t* const VGA_MEMORY = (uint16_t*) 0xB8000;
